+++ template/driver/aws/driver.go
@@ -24,6 +24,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 	"github.com/aws/aws-sdk-go/service/iam/iamiface"
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
+	"github.com/aws/aws-sdk-go/service/sns/snsiface"
 	"github.com/wallix/awless/logger"
 	"github.com/wallix/awless/template/driver"
 )
@@ -32,16 +33,18 @@ type AwsDriver struct {
 	iam iamiface.IAMAPI
 	ec2 ec2iface.EC2API
 	s3  s3iface.S3API
+	sns snsiface.SNSAPI
 
 	dryRun bool
 	logger *logger.Logger
 }
 
-func NewDriver(ec2, iam, s3 interface{}) *AwsDriver {
+func NewDriver(ec2, iam, s3, sns interface{}) *AwsDriver {
 	return &AwsDriver{
 		ec2:    ec2.(ec2iface.EC2API),
 		iam:    iam.(iamiface.IAMAPI),
 		s3:     s3.(s3iface.S3API),
+		sns:    sns.(snsiface.SNSAPI),
 		logger: logger.DiscardLogger,
 	}
 }