+++ template/driver/aws/driver_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 	"github.com/aws/aws-sdk-go/service/iam/iamiface"
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
+	"github.com/aws/aws-sdk-go/service/sns/snsiface"
 )
 
 func TestHumanizeString(t *testing.T) {
@@ -46,7 +47,7 @@ func TestHumanizeString(t *testing.T) {
 
 func TestDriver(t *testing.T) {
 	awsMock := &mockEc2{}
-	driv := NewDriver(awsMock, &mockIam{}, &mockS3{})
+	driv := NewDriver(awsMock, &mockIam{}, &mockS3{}, &mockSNS{})
 
 	t.Run("Create vpc", func(t *testing.T) {
 		cidr := "10.0.0.0/16"
@@ -231,6 +232,10 @@ type mockIam struct {
 
 type mockS3 struct {
 	s3iface.S3API
+}
+
+type mockSNS struct {
+	snsiface.SNSAPI
 }
 
 type mockEc2 struct {