+++ template/driver/generators/aws.go
@@ -136,6 +136,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/iam"
 	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go/service/sns"
 )
 {{ range $index, $def := . }}
 {{- if not $def.ManualFuncDefinition }}